Remove :: from shiboken include path (#48)
The Variable QT_INCLUDE_DIR_WITH_COLONS is iteratively generated
starting with an empty string and adding :${dir}. This always results in
a colon as the first symbol. When combining it with the INCLUDE_PATH
adding a second colon results in :: in the string. The shiboken command
line utility currently has a bug generating an invalid command out of
this (-I without a string). This seems to have been fixed in the
development version already. Still this double colon ins superfluous so
let's remove it.