Use PyQt5 module path to find SIP bindings (ROS 2) (#106)
* Use PyQt5 module path to find SIP bindings (#105)

`sipconfig._pkg_config['default_mod_dir']` is currently used to find the PyQt5 SIP bindings, but this location is determined by where SIP is installed, which may not be the same as where PyQt5 is installed. A real world example of this is in Nix, where each package is installed to a separate isolated directory. Instead, we can use `PyQt5.__path__[0]`, which will always point to the location of the PyQt5 module.

Co-authored-by: Akash Patel <17132214+acxz@users.noreply.github.com>